+++ src/model/ClusterSettings.hh
@@ -2,8 +2,8 @@
 #define __CLUSTER_SETTINGS_HH__
 
 typedef struct ClusterSettings {
-    int M; // number of storage nodes
-    int N; // number of stripes
+    size_t M; // number of storage nodes
+    size_t N; // number of stripes
 } ClusterSettings;
 
 #endif // __CLUSTER_SETTINGS_HH__
