+++ src/model/StripeBatch.hh
@@ -12,17 +12,17 @@ private:
     ConvertibleCode _code;
     ClusterSettings _settings;
     vector<StripeGroup> _stripe_groups;
-    int _id;
+    size_t _id;
 
 public:
-    StripeBatch(int id, ConvertibleCode &code, ClusterSettings &settings);
+    StripeBatch(size_t id, ConvertibleCode &code, ClusterSettings &settings);
     ~StripeBatch();
 
     bool constructInSequence(vector<Stripe> &stripes);
     bool constructByRandomPick(vector<Stripe> &stripes, mt19937 &random_generator);
     bool constructByCost(vector<Stripe> &stripes);
 
-    int getId();
+    size_t getId();
     ConvertibleCode &getCode();
     ClusterSettings &getClusterSettings();
     vector<StripeGroup> &getStripeGroups();
